requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
google-generativeai==0.3.2
google-api-python-client==2.108.0
